refactor(mobile): keep device identity paths within line limits
